namespace UploaderDeamon project and add reference to ProfilerGUI
