+++ Cqse.Teamscale.Profiler.Dotnet.sln
@@ -12,6 +12,9 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UploadDaemon_Test", "UploadDaemon_Test\UploadDaemon_Test.csproj", "{1405B69B-D9A6-47F6-9C05-8B839DB3A0E4}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ProfilerGUI", "ProfilerGUI\ProfilerGUI.csproj", "{0BB237CF-2C22-48DC-95E3-E9F58AB6A398}"
+	ProjectSection(ProjectDependencies) = postProject
+		{A01A7519-E10D-4153-8CD4-678935721DF3} = {A01A7519-E10D-4153-8CD4-678935721DF3}
+	EndProjectSection
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
