+++ ProfilerGUI/packages.config
@@ -3,5 +3,5 @@
   <package id="Newtonsoft.Json" version="11.0.1" targetFramework="net45" />
   <package id="NLog" version="4.5.9" targetFramework="net45" />
   <package id="NLog.Wpf.RichTextBox" version="1.0.2" targetFramework="net45" />
-  <package id="System.ValueTuple" version="4.5.0" targetFramework="net45" />
+  <package id="System.ValueTuple" version="4.5.0" targetFramework="net461" />
 </packages>
