+++ UploadDaemon/UploadDaemon.csproj
@@ -64,9 +64,9 @@
     <Compile Include="TraceFileScanner.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="TeamscaleServer.cs" />
-    <Compile Include="upload\IUpload.cs" />
-    <Compile Include="upload\FileSystemUpload.cs" />
-    <Compile Include="upload\TeamscaleUpload.cs" />
+    <Compile Include="Upload\IUpload.cs" />
+    <Compile Include="Upload\FileSystemUpload.cs" />
+    <Compile Include="Upload\TeamscaleUpload.cs" />
     <Compile Include="Uploader.cs" />
   </ItemGroup>
   <ItemGroup>
